share types with epic
