+++ src/web/components/SlotBodyEnd.tsx
@@ -14,6 +14,7 @@ import {
     getLastOneOffContributionDate,
 } from '@root/src/web/lib/contributions';
 import { initPerf } from '@root/src/web/browser/initPerf';
+import {sendOphanContributionsComponentEvent, TestMeta} from "@root/src/web/browser/ophan/ophan";
 import { getCookie } from '../browser/cookie';
 import { useHasBeenSeen } from '../lib/useHasBeenSeen';
 
@@ -29,33 +30,6 @@ const checkForErrors = (response: any) => {
         );
     }
     return response;
-};
-
-type OphanAction = 'INSERT' | 'VIEW';
-
-type TestMeta = {
-    abTestName: string;
-    abTestVariant: string;
-    campaignCode: string;
-    campaignId: string;
-};
-
-const sendOphanEpicEvent = (action: OphanAction, testMeta: TestMeta): void => {
-    const componentEvent = {
-        component: {
-            componentType: 'ACQUISITIONS_EPIC',
-            products: ['CONTRIBUTION', 'MEMBERSHIP_SUPPORTER'],
-            campaignCode: testMeta.campaignCode,
-            id: testMeta.campaignId,
-        },
-        abTest: {
-            name: testMeta.abTestName,
-            variant: testMeta.abTestVariant,
-        },
-        action,
-    };
-
-    window.guardian.ophan.record({ componentEvent });
 };
 
 const sendOphanReminderEvent = (componentId: string): void => {
@@ -205,7 +179,7 @@ const MemoisedInner = ({
                             onReminderOpen: sendOphanReminderOpenEvent,
                         });
                         setEpic(() => epicModule.ContributionsEpic); // useState requires functions to be wrapped
-                        sendOphanEpicEvent('INSERT', meta);
+                        sendOphanContributionsComponentEvent('INSERT', meta, 'ACQUISITIONS_EPIC');
                     })
                     // eslint-disable-next-line no-console
                     .catch(error => console.log(`epic - error is: ${error}`));
@@ -217,7 +191,7 @@ const MemoisedInner = ({
     useEffect(() => {
         if (hasBeenSeen && epicMeta) {
             logView(epicMeta.abTestName);
-            sendOphanEpicEvent('VIEW', epicMeta);
+            sendOphanContributionsComponentEvent('VIEW', epicMeta, 'ACQUISITIONS_EPIC');
         }
     }, [hasBeenSeen, epicMeta]);
